Move EdgeDegreesService into an analysis subdir
